feat-#24: move cars repository to db module
