+++ src/catalog/cars/commands/handlers/update-cars.handler.ts
@@ -2,7 +2,7 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { CarsRepository } from '../../repositories';
+import { CarsRepository } from '../../../../db/repositories';
 import { UpdateCarsCommand } from '../impl';
 import { CarsEntity } from '../../../../db/entities/cars.entity';
