+++ src/catalog/cars/queries/handlers/get-single-car.handler.ts
@@ -2,7 +2,7 @@ import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { CarsRepository } from '../../repositories';
+import { CarsRepository } from '../../../../db/repositories';
 import { GetSingleCarQuery } from '../impl';
 import { CarsEntity } from '../../../../db/entities/cars.entity';
