--- src/catalog/cars/repositories/index.ts
@@ -1,1 +1,0 @@
-export * from './cars.repository';
