+++ src/db/repositories/cars.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 
-import { CarsEntity } from '../../../db/entities/cars.entity';
+import { CarsEntity } from '../entities/cars.entity';
 
 @EntityRepository(CarsEntity)
 export class CarsRepository extends Repository<CarsEntity> {}
