+++ src/db/repositories/index.ts
@@ -1,1 +1,2 @@
 export * from './activity.repository';
+export * from './cars.repository';
